subscribe on the groups$ observable in schedule.ts
My guess: Looks like the only subscriber on observable was the one that closed the loader. and since it filtered away everything except the first value, the poor observable got forgotten when groups became empty and the whole process stopped working. 
I tried to subscribe on the whole thing, and as well changed closing of the loader based on what confDataService returns. Looks like it works? :)